consolidate loki-http-client into 1 file
