+++ lib/rails_loki_exporter.rb
@@ -7,23 +7,12 @@ module RailsLokiExporter
   class << self
     def create_logger(config_file_path)
       config = load_config(config_file_path)
-
-      connection_instance =  MyConnection.new(
-        config['base_url'],
-        config['user_name'],
-        config['password'],
-        config['auth_enabled'],
-        config['host_name'],
-        config['job_name']
-      )
-
-      client = Client.new(config)
+      client = LokiHttpClient.new(config)
       logger = InterceptingLogger.new(intercept_logs: config['intercept_logs'])
       if config['enable_log_subscriber'] 
         CustomLogSubscriber.client = client
       end
       logger.client = client
-      client.connection = connection_instance
       logger
     end
