--- lib/rails_loki_exporter/my_connection.rb
@@ -1,8 +1,0 @@
-module RailsLokiExporter
-    class MyConnection
-      include Connection
-      def self.create(base_url, user_name, password, auth_enabled, host_name, job_name)
-        new(base_url, user_name, password, auth_enabled, host_name, job_name).connection
-      end
-    end
-end
